refactor: extract tool_dispatch.rs + prompt.rs from inference.rs (#116 P0-2)
inference.rs was doing 5 jobs in 1,324 lines. Extracted:

- tool_dispatch.rs (643 lines): can_parallelize, execute_one_tool,
  execute_tools_parallel, execute_tools_sequential, execute_sub_agent,
  request_approval + 4 tests
- prompt.rs (143 lines): build_system_prompt, list_available_agents
  + 4 tests

inference.rs: 1,324 → 464 lines — now purely the stream-collect
loop + iteration control.

agent.rs updated to import build_system_prompt from prompt module.

All tests pass, clippy clean.
