+++ koda-core/src/lib.rs
@@ -20,8 +20,10 @@ pub mod loop_guard;
 pub mod mcp;
 pub mod memory;
 pub mod preview;
+pub mod prompt;
 pub mod providers;
 pub mod runtime_env;
 pub mod session;
+pub mod tool_dispatch;
 pub mod tools;
 pub mod version;
